replace all alert with sweetalert

+++ src/components/MyReview.jsx
@@ -2,6 +2,8 @@ import { useContext, useState, useEffect } from "react";
 import { useLoaderData } from "react-router-dom";
 import { AuthContext } from "./AuthProvider";
 import { Zoom  } from "react-awesome-reveal";
+import Swal from 'sweetalert2';
+
 
 
 const MyReview = () => {
@@ -20,17 +22,36 @@ const MyReview = () => {
   }, [data, currentUserEmail]);
 
   const handleDelete = (id) => {
-    if (id) {
-      fetch(`http://localhost:5000/all_reviews/${id}`, {
-        method: "DELETE",
-      })
-        .then((res) => res.json())
-        .then((data) => {
-          alert("Review deleted successfully!");
-          const updatedReviews = myReview.filter((review) => id != review._id);
-          setMyReview(updatedReviews);
+
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success"
         });
-    }
+        if (id) {
+          fetch(`http://localhost:5000/all_reviews/${id}`, {
+            method: "DELETE",
+          })
+            .then((res) => res.json())
+            .then((data) => {          
+              const updatedReviews = myReview.filter((review) => id != review._id);
+              setMyReview(updatedReviews);
+            });
+        }
+      }
+    });
+
+   
   };
 
   const handleUpdate = (review) => {
@@ -55,7 +76,10 @@ const MyReview = () => {
     })
       .then((res) => res.json())
       .then(() => {
-        alert("Review updated successfully!");
+        Swal.fire({
+          title: "Review updated successfully!",
+          icon: "success"
+        });
         setSelectedReview(null);
         const updatedReviews = myReview.map((review) =>
           review._id === selectedReview._id
